+++ spec/dummy_helper.rb
@@ -1,0 +1,183 @@
+DUMMY_BATCH = {
+  "default_carrier_account": "903074429eab4954b72df8a70defdfe3",
+  "default_servicelevel_token": "usps_priority",
+  "label_filetype": "ZPLII",
+  "metadata": "BATCH #170",
+  "batch_shipments": [
+    {
+      "shipment": {
+        "address_from": {
+          "name": "Simon Says",
+          "company": "",
+          "street1": "965 Mission St",
+          "street2": "Ste 201",
+          "city": "San Francisco",
+          "state": "CA",
+          "zip": "94103",
+          "country": "US",
+          "phone": "4151234567",
+          "email": "hippo@goshippo.com"
+        },
+        "address_to": {
+          "name": "Who Says",
+          "company": "",
+          "street1": "400 Simon St",
+          "street_no": "",
+          "street2": "Apt 123",
+          "city": "San Francisco",
+          "state": "CA",
+          "zip": "94102",
+          "country": "US",
+          "phone": "4151234567",
+          "email": "hippo@goshippo.com",
+          "metadata": ""
+        },
+        "parcels": [{
+          "length": "5",
+          "width": "5",
+          "height": "5",
+          "distance_unit": "in",
+          "weight": "2",
+          "mass_unit": "oz",
+          "template": "",
+          "metadata": "Customer ID 123456"
+        }],
+        "extra": {
+          "reference_1": "REF1 - Simon",
+          "reference_2": "REF2 - Says",
+        },
+        "metadata": ""
+      }
+    },
+    {
+      "shipment": {
+        "address_from": {
+          "name": "Simon Says",
+          "company": "",
+          "street1": "965 Mission St",
+          "street2": "Ste 201",
+          "city": "San Francisco",
+          "state": "CA",
+          "zip": "94103",
+          "country": "US",
+          "phone": "4151234567",
+          "email": "shippo@goshippo.com"
+        },
+        "address_to": {
+          "name": "Who Says",
+          "company": "",
+          "street1": "400 Simon St",
+          "street_no": "",
+          "street2": "Apt 123",
+          "city": "San Francisco",
+          "state": "CA",
+          "zip": "94102",
+          "country": "US",
+          "phone": "4151234567",
+          "email": "hippo@goshippo.com",
+          "metadata": ""
+        },
+        "parcels": [{
+          "length": "5",
+          "width": "5",
+          "height": "5",
+          "distance_unit": "in",
+          "weight": "2",
+          "mass_unit": "oz",
+          "template": "",
+          "metadata": "Customer ID 123456"
+        }],
+        "extra": {
+          "reference_1": "REF1 - Simon",
+          "reference_2": "REF2 - Says",
+        },
+        "metadata": ""
+      }
+    },
+    {
+      "shipment": {
+        "address_from": {
+          "name": "Simon Says",
+          "company": "",
+          "street1": "965 Mission St",
+          "street2": "Ste 201",
+          "city": "San Francisco",
+          "state": "CA",
+          "zip": "94103",
+          "country": "US",
+          "phone": "4151234567",
+          "email": "hippo@goshippo.com"
+        },
+        "address_to": {
+          "name": "Who Says",
+          "company": "",
+          "street1": "400 Simon St",
+          "street_no": "",
+          "street2": "Apt 123",
+          "city": "San Francisco",
+          "state": "CA",
+          "zip": "94102",
+          "country": "US",
+          "phone": "4151234567",
+          "email": "hippo@goshippo.com",
+          "metadata": ""
+        },
+        "parcels": [{
+          "length": "5",
+          "width": "5",
+          "height": "5",
+          "distance_unit": "in",
+          "weight": "2",
+          "mass_unit": "oz",
+          "template": "",
+          "metadata": "Customer ID 123456"
+        }],
+        "extra": {
+          "reference_1": "REF1 - Simon",
+          "reference_2": "REF2 - Says",
+        },
+        "metadata": ""
+      }
+    }
+  ]
+}
+
+DUMMY_SHIPMENT = {
+    "address_to": {
+        "name": "First Last",
+        "company": "",
+        "street1": "965 Mission St",
+        "street2": "#123",
+        "city": "San Francisco",
+        "state": "CA",
+        "zip": "94103",
+        "country": "US",
+        "phone": "4151234567",
+        "email": "hippo@goshippo.com",
+    },
+    "address_from": {
+        "name": "Amazon Fulfillment Services",
+        "company": "Returns Departments",
+        "street1": "400 Simon St",
+        "street2": "123",
+        "city": "San Francisco",
+        "state": "CA",
+        "zip": "94102",
+        "country": "US",
+        "phone": "4151234567",
+        "email": "hippo@goshippo.com"
+    },
+    "parcels": [{
+        "length": "10",
+        "width": "15",
+        "height": "10",
+        "distance_unit": "in",
+        "weight": "15",
+        "mass_unit": "oz",
+        "metadata": ""
+    }],
+    "extra": {
+      "reference_1": "REF1 - Simon",
+      "reference_2": "REF2 - Says",
+    }
+}