+++ spec/shippo/api/model/parcel_spec.rb
@@ -11,7 +11,6 @@ RSpec.describe 'Shippo::API::Parcel' do
       # test the various options of accessi bg the parcel data
       expect(result[:metadata]).to eql('Customer ID 123456')
       expect(result.object.owner).to eql('tobias.schottdorf@gmail.com')
-      expect(result.object.state.to_s).to match(/valid/i)
     end
 
     context 'should be able to assign and read data' do