+++ CHANGELOG.md
@@ -1,6 +1,42 @@
+#### 3.1.0 release, Sep 11th, 2017
+- Removed deprecated methods
+- Removed hash payload in GET
+- Made RestClient timeouts configurable
+- Suppressed Hashie warnings
+
+#### 3.0.1 release, Jul 14th, 2017
+- Updated rest client dependency
+
+#### 3.0.0 release, Mar 31st, 2017
+- Version 3 changes
+    - Object fields removed/updated
+    - Credentials inserted for tests
+
+#### 2.0.8 release, Feb 15th, 2017
+- Add Batch API with example code
+    - Creation
+    - Retrieval
+    - Adding shipments to Batch objects
+    - Removing shipments from Batch objects
+    - Purchasing
+- Add Track API with example code
+    - Retrieving tracking status of a shipment
+    - Registering a tracking webhook
+- Add mock tests using VCR (for Batch and Track API)
+    - https://github.com/vcr/vcr
+
+#### 2.0.7 release, Jan 2nd, 2017
+- Fixed bug preventing address validation
+- Removed trailing slash from base URL, added spec test to ensure this configuration
+- Added address validation example
+- Updated basic shipment example to include how to access the Shippo object id
+#### 2.0.6 release, Nov 22nd, 2016
+- Fixed bug to send request with correct API version header
+
 #### 2.0.5 release, Oct 24th, 2016
  - Updated README.md
  - now possible to send API version
+ 
 #### 2.0.4 release, Oct 6th 2016
  - Rails5 Compatibility via relaxed dependencies
    - removing mime-types dependency